add a faker library provider for manifest status and use in manifest_factory
